requirements.txt
streamlit==1.40.0
requests==2.31.0
pandas==2.2.0
streamlit-authenticator
google-auth
google-auth-oauthlib